requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
ollama>=0.1.0
numpy>=1.21.0
pydantic>=2.0.0
python-dotenv>=1.0.0
psycopg-pool>=3.1.0
pgvector>=0.2.0